Created Directory Card and moved BitCard to said directory
