+++ src/Components/Cards/BitCard.js
